add BaseMCPKnowledgeSource
